+++ week2/include/lib.h
@@ -1,0 +1,15 @@
+/* Copyright 2021 Carlos Pinto Machado */
+/* See LICENSE for details */
+
+#ifndef WEEK2_INCLUDE_LIB_H_
+#define WEEK2_INCLUDE_LIB_H_
+
+/* FUNCTION DEFINITIONS */
+
+/* func_1: a function within lib */
+void func_1();
+
+/* func_2: a function within lib */
+void func_2();
+
+#endif /* WEEK2_INCLUDE_LIB_H_ */