+++ week2/lib1.c
@@ -1,0 +1,16 @@
+/* Copyright 2021 Carlos Pinto Machado */
+/* See LICENSE for details */
+
+/* HEADERS */
+#include <stdio.h>
+
+
+/* FUNCTION IMPLEMENTATIONS */
+
+void func_1() {
+    printf("func 1 inside lib1\n");
+}
+
+void func_2() {
+    printf("func 2 inside lib2\n");
+}