+++ week2/prog1.c
@@ -1,0 +1,14 @@
+/* Copyright 2021 Carlos Pinto Machado */
+/* See LICENSE for details */
+
+/* HEADERS */
+#include "include/lib.h"
+
+
+/* FUNCTION IMPLEMENTATIONS */
+
+int main(void) {
+    func_1();
+    func_2();
+    return 0;
+}